requirements.txt
deepagents==0.1.2
google-genai==1.45.0
python-dotenv
